requirements.txt
Django==3.0.2
xmltodict==0.12.0
requests==2.22.0
djangorestframework==3.11.0